moved matrix + trust adapter to external folder
