+++ src/domain/agent/agent/agent.module.ts
@@ -6,7 +6,7 @@ import { Agent } from '@domain/agent/agent';
 import { AgentResolverFields } from './agent.resolver.fields';
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
-import { TrustRegistryAdapterModule } from '../../../services/platform/trust-registry/trust.registry.adapter/trust.registry.adapter.module';
+import { TrustRegistryAdapterModule } from '@services/external/trust-registry/trust.registry.adapter/trust.registry.adapter.module';
 import { AgentResolverMutations } from './agent.resolver.mutations';
 import { VerifiedCredentialModule } from '../verified-credential/verified.credential.module';
 import { AgentResolverSubscriptions } from '@domain/agent/agent/agent.resolver.subscriptions';
