+++ src/domain/agent/agent/agent.service.ts
@@ -29,8 +29,8 @@ import { FindOneOptions, Repository } from 'typeorm';
 import { CredentialService } from '../credential/credential.service';
 import { WalletManagerCommand } from '@common/enums/wallet.manager.command';
 import { CredentialMetadataOutput } from '../verified-credential/dto/verified.credential.dto.metadata';
-import { IClaim } from '@services/platform/trust-registry/trust.registry.claim/claim.interface';
-import { TrustRegistryAdapter } from '@services/platform/trust-registry/trust.registry.adapter/trust.registry.adapter';
+import { IClaim } from '@services/external/trust-registry/trust.registry.claim/claim.interface';
+import { TrustRegistryAdapter } from '@services/external/trust-registry/trust.registry.adapter/trust.registry.adapter';
 import { GrantCredentialInput } from './dto/agent.dto.credential.grant';
 import { RevokeCredentialInput } from './dto/agent.dto.credential.revoke';
 import { AgentBeginVerifiedCredentialRequestOutput } from './dto/agent.dto.verified.credential.request.begin.output';
