+++ src/domain/community/community/community.resolver.mutations.ts
@@ -22,9 +22,9 @@ import { ApplicationAuthorizationService } from '../application/application.serv
 import { AgentService } from '@domain/agent/agent/agent.service';
 import { CommunityJoinInput } from './dto/community.dto.join';
 import { CommunityApplyInput } from './dto/community.dto.apply';
-import { CommunityMemberClaim } from '@services/platform/trust-registry/trust.registry.claim/claim.community.member';
+import { CommunityMemberClaim } from '@services/external/trust-registry/trust.registry.claim/claim.community.member';
 import { AgentBeginVerifiedCredentialOfferOutput } from '@domain/agent/agent/dto/agent.dto.verified.credential.offer.begin.output';
-import { AlkemioUserClaim } from '@services/platform/trust-registry/trust.registry.claim/claim.alkemio.user';
+import { AlkemioUserClaim } from '@services/external/trust-registry/trust.registry.claim/claim.alkemio.user';
 import { CreateFeedbackOnCommunityContextInput } from '@domain/community/community/dto/community.dto.create.feedback.on.context';
 import { CreateUserGroupInput } from '../user-group/dto';
 import { AssignCommunityMemberOrganizationInput } from './dto/community.dto.assign.member.organization';
