+++ src/services/external/matrix/adapter-group/matrix.group.adapter.ts
