+++ src/services/external/matrix/adapter-group/matrix.group.dto.result.ts
