+++ src/services/external/matrix/adapter-message/matrix.message.adapter.ts
