+++ src/services/external/matrix/adapter-room/matrix.room.adapter.ts
