+++ src/services/external/matrix/adapter-room/matrix.room.dto.create.options.ts
