+++ src/services/external/matrix/adapter-room/matrix.room.dto.options.ts
