+++ src/services/external/matrix/adapter-room/matrix.room.dto.response.message.ts
