+++ src/services/external/matrix/adapter-room/matrix.room.ts
