+++ src/services/external/matrix/adapter-user/matrix.user.adapter.ts
