+++ src/services/external/matrix/adapter-user/matrix.user.dto.logged.in.response.ts
