+++ src/services/external/matrix/adapter-user/matrix.user.interface.ts
