+++ src/services/external/matrix/agent-pool/matrix.agent.pool.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { MatrixAgentPool } from '@src/services/platform/matrix/agent-pool/matrix.agent.pool';
-import { MatrixUserManagementModule } from '@src/services/platform/matrix/management/matrix.user.management.module';
+import { MatrixAgentPool } from '@services/external/matrix/agent-pool/matrix.agent.pool';
+import { MatrixUserManagementModule } from '@services/external/matrix/management/matrix.user.management.module';
 import { MatrixAgentModule } from '../agent/matrix.agent.module';
 
 @Module({
