+++ src/services/external/matrix/events/matrix.event.adapter.group.ts
