+++ src/services/external/matrix/management/matrix.user.management.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
-import { MatrixCryptographyModule } from '@src/services/platform/matrix/cryptography/matrix.cryptography.module';
-import { MatrixUserManagementService } from '@src/services/platform/matrix/management/matrix.user.management.service';
+import { MatrixCryptographyModule } from '@services/external/matrix/cryptography/matrix.cryptography.module';
+import { MatrixUserManagementService } from '@services/external/matrix/management/matrix.user.management.service';
 import { MatrixUserAdapterModule } from '../adapter-user/matrix.user.adapter.module';
 
 @Module({
