+++ src/services/external/trust-registry/trust.registry.adapter/trust.registry.adapter.module.ts
