+++ src/services/external/trust-registry/trust.registry.claim/claim.community.member.ts
