+++ src/services/external/trust-registry/trust.registry.configuration/trust.registry.configuration.adapter.ts
