+++ src/services/external/trust-registry/trust.registry.configuration/trust.registry.dto.credential.metadata.ts
