+++ src/services/platform/communication-adapter/communication-adapter.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { MatrixAgentPoolModule } from '@src/services/platform/matrix/agent-pool/matrix.agent.pool.module';
-import { MatrixUserManagementModule } from '@src/services/platform/matrix/management/matrix.user.management.module';
-import { MatrixGroupAdapterModule } from '../matrix/adapter-group/matrix.group.adapter.module';
-import { MatrixRoomAdapterModule } from '../matrix/adapter-room/matrix.room.adapter.module';
-import { MatrixAgentModule } from '../matrix/agent/matrix.agent.module';
-import { MatrixUserAdapterModule } from '../matrix/adapter-user/matrix.user.adapter.module';
+import { MatrixAgentPoolModule } from '@services/external/matrix/agent-pool/matrix.agent.pool.module';
+import { MatrixUserManagementModule } from '@services/external/matrix/management/matrix.user.management.module';
+import { MatrixGroupAdapterModule } from '@services/external/matrix/adapter-group/matrix.group.adapter.module';
+import { MatrixRoomAdapterModule } from '@services/external/matrix/adapter-room/matrix.room.adapter.module';
+import { MatrixAgentModule } from '@services/external/matrix/agent/matrix.agent.module';
+import { MatrixUserAdapterModule } from '@services/external/matrix/adapter-user/matrix.user.adapter.module';
 import { CommunicationAdapter } from './communication.adapter';
-import { MicroservicesModule } from '../../../core/microservices/microservices.module';
+import { MicroservicesModule } from '@core/microservices/microservices.module';
 
 @Module({
   imports: [
