+++ src/services/platform/communication-adapter/communication.adapter.ts
@@ -6,16 +6,16 @@ import { CommunicationRoomResult } from '@domain/communication/room/dto/communic
 import { DirectRoomResult } from '@domain/community/user/dto/user.dto.communication.room.direct.result';
 import { Inject, Injectable, LoggerService } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { MatrixAgentPool } from '@src/services/platform/matrix/agent-pool/matrix.agent.pool';
+import { MatrixAgentPool } from '@services/external/matrix/agent-pool/matrix.agent.pool';
 import { MatrixClient } from 'matrix-js-sdk';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
-import { MatrixGroupAdapter } from '../matrix/adapter-group/matrix.group.adapter';
-import { MatrixRoomAdapter } from '../matrix/adapter-room/matrix.room.adapter';
-import { MatrixUserAdapter } from '../matrix/adapter-user/matrix.user.adapter';
-import { IOperationalMatrixUser } from '../matrix/adapter-user/matrix.user.interface';
-import { MatrixAgent } from '../matrix/agent/matrix.agent';
-import { MatrixAgentService } from '../matrix/agent/matrix.agent.service';
-import { MatrixUserManagementService } from '../matrix/management/matrix.user.management.service';
+import { MatrixGroupAdapter } from '@services/external/matrix/adapter-group/matrix.group.adapter';
+import { MatrixRoomAdapter } from '@services/external/matrix/adapter-room/matrix.room.adapter';
+import { MatrixUserAdapter } from '@services/external/matrix/adapter-user/matrix.user.adapter';
+import { IOperationalMatrixUser } from '@services/external/matrix/adapter-user/matrix.user.interface';
+import { MatrixAgent } from '@services/external/matrix/agent/matrix.agent';
+import { MatrixAgentService } from '@services/external/matrix/agent/matrix.agent.service';
+import { MatrixUserManagementService } from '@services/external/matrix/management/matrix.user.management.service';
 import { CommunicationDeleteMessageInput } from './dto/communication.dto.message.delete';
 import { CommunicationEditMessageInput } from './dto/communication.dto.message.edit';
 import { CommunicationSendMessageInput } from './dto/communication.dto.message.send';
