+++ src/services/platform/wallet-manager-adapter/dto/wallet.manager.dto.request.vc.begin.ts
@@ -1,4 +1,4 @@
-import { TrustRegistryCredentialMetadata } from '@services/platform/trust-registry/trust.registry.configuration/trust.registry.dto.credential.metadata';
+import { TrustRegistryCredentialMetadata } from '@services/external/trust-registry/trust.registry.configuration/trust.registry.dto.credential.metadata';
 
 export class WalletManagerRequestVcBegin {
   issuerDID!: string;
