+++ src/services/platform/wallet-manager-adapter/wallet.manager.adapter.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { TrustRegistryAdapterModule } from '../trust-registry/trust.registry.adapter/trust.registry.adapter.module';
+import { TrustRegistryAdapterModule } from '@services/external/trust-registry/trust.registry.adapter/trust.registry.adapter.module';
 import { WalletManagerAdapter } from './wallet.manager.adapter';
 
 @Module({
